Add fish/conf.d linking to setup.sh

+++ setup.sh
@@ -1,8 +1,10 @@
 #!/bin/bash
 
 REPO_DIR=$(dirname $0)
-FILES=$(ls $REPO_DIR/dot.*)
-VIM_PLUG=~/.vim/autoload/plug.vim
+FILES=$(ls -d $REPO_DIR/dot.*)
+VIM_PLUG="$HOME/.vim/autoload/plug.vim"
+FISH_CONF_FROM="$REPO_DIR/dot.config/fish/conf.d"
+FISH_CONF_TO="$HOME/.config/fish/conf.d"
 
 function _linkpath {
   BASENAME=$(basename $1)
@@ -10,13 +12,7 @@ function _linkpath {
   echo "$HOME/$NO_DOT"
 }
 
-function _link_it {
-  for f in $FILES; do
-    LINKPATH=$(_linkpath $f)
-    echo -n "Linking $f to $LINKPATH : "
-    ln -s $f $LINKPATH && echo "done"
-  done
-
+function _install_vim_plug {
   if [ ! -f $VIM_PLUG ]; then
     echo "Install vim-plug..."
     # https://github.com/junegunn/vim-plug
@@ -24,18 +20,51 @@ function _link_it {
   fi
 }
 
-function _clean_it {
+function _link_fish_conf {
+  echo -n "Linking $FISH_CONF_FROM directory : "
+
+  if [ -d $FISH_CONF_TO ]; then
+    echo "directory $FISH_CONF_TO already exists, consider to back it up!"
+    exit 1
+  fi
+
+  ln -s $FISH_CONF_FROM $FISH_CONF_TO && echo "done"
+}
+
+function _unlink_fish_conf {
+  echo -n "Removing $FISH_CONF_FROM directory : "
+  rm $FISH_CONF_TO && echo "done"
+}
+
+function _create_links {
   for f in $FILES; do
-    LINKPATH=$(_linkpath $f)
-    echo -n "Removing $LINKPATH : "
-    rm $LINKPATH && echo "done"
+    if [ ! -d $f ]; then
+      LINKPATH=$(_linkpath $f)
+      echo -n "Linking $f to $LINKPATH : "
+      ln -s $f $LINKPATH && echo "done"
+    fi
   done
+
+  _install_vim_plug
+  _link_fish_conf
+}
+
+function _remove_links {
+  for f in $FILES; do
+    if [ ! -d $f ]; then
+      LINKPATH=$(_linkpath $f)
+      echo -n "Removing $LINKPATH : "
+      rm $LINKPATH && echo "done"
+    fi
+  done
+
+  _unlink_fish_conf
 }
 
 while [ -n "$1" ]; do
   case "$1" in
-    clean) _clean_it;;
-    link) _link_it;;
+    link) _create_links;;
+    clean) _remove_links;;
   esac
   shift
 done
